+++ tests/FunctionsTest.php
@@ -35,15 +35,19 @@ class FunctionsTest extends \PHPUnit_Framework_TestCase
     public function testUris()
     {
         return [
+            // Technically the spec adds the superfluous # at the end, but we don't need to enforce that.
             ['http://x.y.z/rootschema.json#', '', 'http://x.y.z/rootschema.json#'],
             ['#foo', 'http://x.y.z/rootschema.json#', 'http://x.y.z/rootschema.json#foo'],
-            // Technically the spec adds the superfluous # at the end, but we don't need to enforce that.
             ['otherschema.json', 'http://x.y.z/rootschema.json#', 'http://x.y.z/otherschema.json'],
             ['#bar', 'http://x.y.z/otherschema.json#', 'http://x.y.z/otherschema.json#bar'],
             ['t/inner.json#a', 'http://x.y.z/otherschema.json#', 'http://x.y.z/t/inner.json#a'],
-            ['some://where.else/completely#', 'http://x.y.z/rootschema.json#', 'some://where.else/completely#'],
+            ['some://where.else/completely#', 'http://x.y.z/rootschema.json#', 'some://where.else/completely'],
             ['folderInteger.json', 'http://localhost:1234/folder/', 'http://localhost:1234/folder/folderInteger.json'],
             ['some-id.json', '', 'some-id.json'],
+            ['item.json', 'http://some/where/other-item.json', 'http://some/where/item.json'],
+            // @todo: fixed in https://github.com/fruux/sabre-uri/pull/10,
+            // we just need to wait for a PHP 5.5 compatible release to update composer.
+            // ['item.json', 'file:///schemas/other-item.json', 'file:///schemas/item.json'],
         ];
     }